feat(consola): implementar consola en segundo plano y comandos de depuracion
- Arquitectura de comandos (GestorComandos, Comando) basada en cola concurrente Lock-Free integrada a 60 APS.
- Lector de System.in en hilo Daemon en segundo plano sin impacto de CPU ni bloqueo de la JVM.
- Comandos añadidos: curar, clima, tp, hora, luz, spawnluz, ambiente, sigilo, killall y velocidad.
- Fix: correccion de recursion mutua (StackOverflowError) en desvinculacion bidireccional entre FuenteLuz y Ente.

+++ src/principal/entes/criaturas/Jugador.java
@@ -826,7 +826,7 @@ public class Jugador extends Criatura {
 		return (int) Math.round(this.desplazamientoY);
 	}
 
-	public void establecerPosicion(final int x, final int y) {
+	public void establecerPosicion(final double x, final double y) {
 		this.x = x;
 		this.y = y;
 		this.moviendoPorRecorrido = false;
